requirements.txt
aniso8601==9.0.1
attrs==21.4.0
click==8.1.3
flasgger==0.9.5
Flask==2.1.3
Flask-JWT-Extended==4.4.2
flask-restplus==0.13.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.7.2
MarkupSafe==2.1.1
mistune==2.0.4
psycopg2-binary==2.9.3
PyJWT==2.4.0
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.39
Werkzeug==2.1.2
zipp==3.8.1
